+++ api/v1alpha1/scaledobject_types.go
@@ -29,6 +29,25 @@ type ScaledObject struct {
 	Status ScaledObjectStatus `json:"status,omitempty"`
 }
 
+// HealthStatus is the status for a ScaledObject's health
+type HealthStatus struct {
+	// +optional
+	NumberOfFailures *int32 `json:"numberOfFailures,omitempty"`
+	// +optional
+	Status HealthStatusType `json:"status,omitempty"`
+}
+
+// HealthStatusType is an indication of whether the health status is happy or failing
+type HealthStatusType string
+
+const (
+	// HealthStatusHappy means the status of the health object is happy
+	HealthStatusHappy HealthStatusType = "Happy"
+
+	// HealthStatusFailing means the status of the health object is failing
+	HealthStatusFailing HealthStatusType = "Failing"
+)
+
 // ScaledObjectSpec is the spec for a ScaledObject resource
 type ScaledObjectSpec struct {
 	ScaleTargetRef *ScaleTarget `json:"scaleTargetRef"`
@@ -44,6 +63,14 @@ type ScaledObjectSpec struct {
 	Advanced *AdvancedConfig `json:"advanced,omitempty"`
 
 	Triggers []ScaleTriggers `json:"triggers"`
+	// +optional
+	Fallback *Fallback `json:"fallback,omitempty"`
+}
+
+// Fallback is the spec for fallback options
+type Fallback struct {
+	FailureThreshold int32 `json:"failureThreshold"`
+	Replicas         int32 `json:"replicas"`
 }
 
 // AdvancedConfig specifies advance scaling options
@@ -79,7 +106,11 @@ type ScaleTriggers struct {
 	Metadata map[string]string `json:"metadata"`
 	// +optional
 	AuthenticationRef *ScaledObjectAuthRef `json:"authenticationRef,omitempty"`
+	// +optional
+	FallbackReplicas *int32 `json:"fallback,omitempty"`
 }
+
+// +k8s:openapi-gen=true
 
 // ScaledObjectStatus is the status for a ScaledObject resource
 // +optional
@@ -98,6 +129,8 @@ type ScaledObjectStatus struct {
 	ResourceMetricNames []string `json:"resourceMetricNames,omitempty"`
 	// +optional
 	Conditions Conditions `json:"conditions,omitempty"`
+	// +optional
+	Health map[string]HealthStatus `json:"health,omitempty"`
 }
 
 // +kubebuilder:object:root=true